pyproject.toml
[project]
name = "aws-cost-negotiator"
version = "0.1.0"
description = "A ReAct agent that analyzes AWS bills and negotiates for cost credits"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
